+++ librustzcash/Cargo.toml
@@ -24,6 +24,5 @@ lazy_static = "1"
 byteorder = "1"
 rand_core = "0.5"
 rand_os = "0.2"
-sapling-crypto = { path = "../sapling-crypto" }
 zcash_primitives = { path = "../zcash_primitives" }
 zcash_proofs = { path = "../zcash_proofs" }
