+++ zcash_client_backend/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 [dependencies]
 bech32 = "0.7"
 pairing = { path = "../pairing" }
-sapling-crypto = { path = "../sapling-crypto" }
 zcash_primitives = { path = "../zcash_primitives" }
 
 [dev-dependencies]
