+++ zcash_primitives/Cargo.toml
@@ -8,6 +8,7 @@ authors = [
 [dependencies]
 aes = "0.3"
 blake2b_simd = "0.5"
+blake2s_simd = "0.5"
 byteorder = "1"
 crypto_api_chachapoly = "0.2.1"
 ff = { path = "../ff" }
@@ -18,8 +19,8 @@ pairing = { path = "../pairing" }
 rand = "0.7"
 rand_core = "0.5"
 rand_os = "0.2"
-sapling-crypto = { path = "../sapling-crypto" }
 sha2 = "0.8"
 
 [dev-dependencies]
+hex-literal = "0.1"
 rand_xorshift = "0.2"
