+++ zcash_primitives/src/transaction/tests.rs
@@ -1,7 +1,8 @@
 use ff::Field;
 use pairing::bls12_381::Bls12;
 use rand_os::OsRng;
-use sapling_crypto::jubjub::{fs::Fs, FixedGenerators};
+
+use crate::jubjub::{fs::Fs, FixedGenerators};
 
 use super::{components::Amount, sighash::signature_hash, Transaction, TransactionData};
 use legacy::Script;
