+++ zcash_primitives/src/util.rs
@@ -1,5 +1,6 @@
 use blake2b_simd::Params;
-use sapling_crypto::jubjub::{JubjubEngine, ToUniform};
+
+use crate::jubjub::{JubjubEngine, ToUniform};
 
 pub fn hash_to_scalar<E: JubjubEngine>(persona: &[u8], a: &[u8], b: &[u8]) -> E::Fs {
     let mut hasher = Params::new().hash_length(64).personal(persona).to_state();
