+++ zcash_proofs/Cargo.toml
@@ -13,7 +13,6 @@ directories = { version = "1", optional = true }
 ff = { path = "../ff" }
 pairing = { path = "../pairing" }
 rand_os = "0.2"
-sapling-crypto = { path = "../sapling-crypto" }
 zcash_primitives = { path = "../zcash_primitives" }
 
 [dev-dependencies]
