+++ zcash_proofs/src/sapling/mod.rs
@@ -1,5 +1,5 @@
 use pairing::bls12_381::Bls12;
-use sapling_crypto::jubjub::{
+use zcash_primitives::jubjub::{
     edwards, fs::FsRepr, FixedGenerators, JubjubBls12, JubjubParams, Unknown,
 };
 use zcash_primitives::transaction::components::Amount;
